Rename HasTappableComponents and Draggable game mixins
