+++ packages/flame/lib/src/events/flame_game_mixins/multi_tap_dispatcher.dart
