+++ packages/flame/lib/src/events/tagged_component.dart
@@ -1,6 +1,6 @@
 import 'package:flame/src/components/core/component.dart';
-import 'package:flame/src/events/flame_game_mixins/has_draggable_components.dart';
-import 'package:flame/src/events/flame_game_mixins/has_tappable_components.dart';
+import 'package:flame/src/events/flame_game_mixins/multi_drag_dispatcher.dart';
+import 'package:flame/src/events/flame_game_mixins/multi_tap_dispatcher.dart';
 import 'package:meta/meta.dart';
 
 /// [TaggedComponent] is a utility class that represents a pair of a component
