+++ packages/flame/test/components/input/sprite_button_component_test.dart
@@ -4,7 +4,7 @@ import 'package:flame/game.dart';
 import 'package:flame/input.dart';
 import 'package:flame/src/anchor.dart';
 import 'package:flame/src/components/sprite_group_component.dart';
-import 'package:flame/src/events/flame_game_mixins/has_tappable_components.dart';
+import 'package:flame/src/events/flame_game_mixins/multi_tap_dispatcher.dart';
 import 'package:flame/src/spritesheet.dart';
 import 'package:flame_test/flame_test.dart';
 import 'package:flutter/material.dart';
