+++ packages/flame/test/events/flame_game_mixins/has_draggable_components_test.dart
@@ -3,7 +3,7 @@
 import 'package:flame/components.dart';
 import 'package:flame/events.dart';
 import 'package:flame/game.dart';
-import 'package:flame/src/events/flame_game_mixins/has_draggable_components.dart';
+import 'package:flame/src/events/flame_game_mixins/multi_drag_dispatcher.dart';
 import 'package:flutter_test/flutter_test.dart';
 
 void main() {
